authapp/requirements.txt
alembic==1.8.0
aniso8601==9.0.1
asgiref==3.5.2
async-timeout==4.0.2
attrs==21.4.0
Authlib==1.0.1
autopep8==1.6.0
Babel==2.10.1
blinker==1.4
cachelib==0.8.0
certifi==2022.6.15
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
commonmark==0.9.1
cryptography==37.0.2
Deprecated==1.2.13
Flask==2.1.2
Flask-BabelEx==0.9.4
Flask-JWT-Extended==4.4.1
Flask-Limiter==2.4.6
Flask-Login==0.6.1
Flask-Mail==0.9.1
Flask-Migrate==3.1.0
Flask-Principal==0.4.0
flask-restx==0.5.1
Flask-Security==3.0.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
gevent==21.12.0
googleapis-common-protos==1.56.4
greenlet==1.1.2
grpcio==1.47.0
idna==3.3
importlib-metadata==4.11.4
importlib-resources==5.7.1
iniconfig==1.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.6.0
limits==2.7.0
Mako==1.2.0
MarkupSafe==2.1.1
marshmallow==3.16.0
multidict==6.0.2
oauthlib==2.1.0
opentelemetry-api==1.10.0
opentelemetry-exporter-jaeger==1.10.0
opentelemetry-exporter-jaeger-proto-grpc==1.10.0
opentelemetry-exporter-jaeger-thrift==1.10.0
opentelemetry-instrumentation==0.29b0
opentelemetry-instrumentation-flask==0.29b1
opentelemetry-instrumentation-wsgi==0.29b0
opentelemetry-sdk==1.10.0
opentelemetry-semantic-conventions==0.29b0
opentelemetry-util-http==0.29b0
packaging==21.3
passlib==1.7.4
pluggy==1.0.0
protobuf==4.21.4
psycopg2-binary==2.9.3
py==1.11.0
pycodestyle==2.8.0
pycparser==2.21
pydantic==1.9.1
Pygments==2.12.0
PyJWT==2.4.0
pyparsing==3.0.9
pyrsistent==0.18.1
pytest==7.1.2
pytest-asyncio==0.18.3
python-logstash==0.4.8
pytz==2022.1
redis==4.3.3
requests==2.28.0
requests-oauthlib==1.1.0
rich==12.5.1
sentry-sdk==1.9.2
six==1.16.0
speaklater==1.3
SQLAlchemy==1.4.37
thrift==0.16.0
toml==0.10.2
tomli==2.0.1
typing-extensions==4.2.0
urllib3==1.26.11
uWSGI==2.0.20
Werkzeug==2.1.2
wrapt==1.14.1
WTForms==3.0.1
zipp==3.8.0
zope.event==4.5.0
zope.interface==5.4.0
